pyproject.toml
[project]
name = "sanaap-backend-challenge-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "django>=5.2.8",
    "django-auditlog>=3.3.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-minio-storage>=0.5.8",
    "django-redis>=6.0.0",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "isort>=7.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
]
